Backend/requirements.txt
# base.txt
asgiref==3.5.2
Django==4.1.5
djangorestframework==3.14.0
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.6
django-environ==0.9.0
django-debug-toolbar
python-dotenv
requests==2.28.1
epydoc
djangorestframework-simplejwt
drf-yasg==1.21.4
django-rest-swagger==2.2.0
coreapi==2.3.3
zulu==2.0.0
djangorestframework-camel-case
pandas == 1.5.2
djangorestframework-dataclasses
django-cors-headers
python-dateutil
